APIs

MatMulAlgorithms.StrassenStage — Type

StrassenStage{T}

A pre-allocated buffer for one stage of Strassen multiplication for eltype T.

Let the multiplication be Y = A * B where Y is M × N, A is M × K and B is K × N. In Strassen's algorithm, all 3 matrices are evenly partitioned into 4 blocks. If some of the dimensions are odd, the last row/column is handled separately, and are flagged by mtail, ktail and ntail.

Fields

mtail, ktail, ntail::Bool.

m, k, n::Int: sizes of partitioned blocks. M = 2m + mtail and so on.

TY, TA, TB::Matrix{T}: Buffers. TY is m × k, TA is k × n, TB is m × n.

MatMulAlgorithms.gaius_mul! — Method

gaius_mul!(C, A, B; multithreaded=true)

Performs C = A*B. This is a wrapper around Gaius.jl.

MatMulAlgorithms.octavian_mul! — Method

octavian_mul!(C, A, B)

Performs C = A*B. This is a wrapper around Octavian.jl.

MatMulAlgorithms.stevengj_mul! — Method

stevengj_mul!(C, A, B; threshold=256)

Performs C = A*B + C, using divide-and-conquer, and LoopVectorization.jl. Reference: https://discourse.julialang.org/t/julia-matrix-multiplication-performance/55175/12

MatMulAlgorithms.strassen_plan — Method

strassen_plan(T, m, k, n; threshold = 64)

Generates a vector of StrassenStages for the Strassen multiplication of m × k and k × n matrices of eltype T. The threshold determines the minimum size of the matrices that are multiplied directly.

MatMulAlgorithms.winograd_mul! — Method

winograd_mul!(Y, A, B; plan::Vector{StrassenStage{T}}, stageind) where T

Performs Y = A*B + Y using Winograd's modification of Strassen's algorithm. plan and stageind are optional. When the arguments have small sizes, the fallback method is winograd_mul_fallback! which by default redirect to LinearAlgebra._mul!.

Note

This method is not numerically accurate. Use it only for exact algebra.

MatMulAlgorithms.winograd_mul_fallback! — Method

winograd_mul_fallback!(Y, A, B)

The fallback method of winograd_mul!. Override this for better tuning.